% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{accessors}
\alias{accessors}
\alias{getProductionScores,scMetaLink-method}
\alias{getSensingScores}
\alias{getSensingScores,scMetaLink-method}
\alias{getCommunicationScores}
\alias{getCommunicationScores,scMetaLink-method}
\alias{getSignificantInteractions}
\alias{getSignificantInteractions,scMetaLink-method}
\alias{getPathwayAggregated}
\alias{getPathwayAggregated,scMetaLink-method}
\alias{getParameters}
\alias{getParameters,scMetaLink-method}
\title{Accessor Functions}
\usage{
getProductionScores(object)

\S4method{getProductionScores}{scMetaLink}(object)

getSensingScores(object)

\S4method{getSensingScores}{scMetaLink}(object)

getCommunicationScores(object)

\S4method{getCommunicationScores}{scMetaLink}(object)

getSignificantInteractions(object)

\S4method{getSignificantInteractions}{scMetaLink}(object)

getPathwayAggregated(object)

\S4method{getPathwayAggregated}{scMetaLink}(object)

getParameters(object)

\S4method{getParameters}{scMetaLink}(object)
}
\arguments{
\item{object}{scMetaLink object}
}
\description{
Accessor Functions
}
