% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactate.R
\name{checkLactateGenes}
\alias{checkLactateGenes}
\title{Check Lactate Gene Availability}
\usage{
checkLactateGenes(object)
}
\arguments{
\item{object}{scMetaLink object}
}
\value{
data.frame showing gene availability by category
}
\description{
Checks which lactate signaling genes are available in the expression data.
Useful for quality control before running analysis.
}
\examples{
\donttest{
data(crc_example)
obj <- createScMetaLink(crc_expr, crc_meta, "cell_type")
checkLactateGenes(obj)
}
}
