% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{citationScMetaLink}
\alias{citationScMetaLink}
\title{Show Package Citation}
\usage{
citationScMetaLink()
}
\value{
Invisibly returns NULL, prints citation information
}
\description{
Display the recommended citation for scMetaLink
}
\examples{
citationScMetaLink()
}
