% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crc_expr}
\alias{crc_expr}
\title{CRC Example Expression Data}
\format{
A sparse dgCMatrix with 4,210 genes (rows) x 2,850 cells (columns).
  Only genes present in MetalinksDB are included to reduce file size.
}
\source{
CellScope package example data
}
\usage{
crc_expr
}
\description{
Example single-cell RNA-seq expression matrix from colorectal cancer
}
\details{
This is a subset of colorectal cancer single-cell data containing:
\itemize{
  \item Tumor cells: Tumor Epithelial (600 cells)
  \item Immune cells: T (500), Plasma (250), B (150), TAM (150), Monocyte (120),
        Normal Macrophage (150), Mast (30)
  \item Stromal cells: CAF (200), Normal Fibroblast (200), Endothelial (100),
        Pericyte (50), SMC (30)
  \item Epithelial: Normal Epithelial (300)
  \item Other: Gliacyte (20)
}
}
\examples{
\donttest{
data(crc_expr)
data(crc_meta)

# Run scMetaLink analysis
obj <- createScMetaLink(crc_expr, crc_meta, "cell_type")
}
}
\keyword{datasets}
