% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.clean_protein_type}
\alias{.clean_protein_type}
\title{Clean protein_type field (remove extra quotes)}
\usage{
.clean_protein_type(type_str)
}
\description{
Clean protein_type field (remove extra quotes)
}
\keyword{internal}
