% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{.compute_spatial_communication_region}
\alias{.compute_spatial_communication_region}
\title{Compute Region-Level Spatial Communication}
\usage{
.compute_spatial_communication_region(
  prod_scores,
  sens_scores,
  weight_matrix,
  cell_meta,
  cell_type_col,
  cell_types,
  common_mets,
  comm_method,
  min_production,
  min_sensing,
  verbose
)
}
\description{
Compute Region-Level Spatial Communication
}
\keyword{internal}
