% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_colors}
\alias{.get_colors}
\title{Get Color Palette}
\usage{
.get_colors(n, palette = "default")
}
\arguments{
\item{n}{Number of colors}

\item{palette}{Character. Palette name}
}
\description{
Get Color Palette
}
\keyword{internal}
