% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.get_degradation_enzymes}
\alias{.get_degradation_enzymes}
\title{Get Degradation Enzymes for Metabolites}
\usage{
.get_degradation_enzymes(db)
}
\arguments{
\item{db}{MetalinksDB database list}
}
\value{
data.frame of degradation relationships
}
\description{
Get Degradation Enzymes for Metabolites
}
\keyword{internal}
