% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.get_extracellular_metabolites}
\alias{.get_extracellular_metabolites}
\title{Get Extracellular Metabolites}
\usage{
.get_extracellular_metabolites(db)
}
\arguments{
\item{db}{MetalinksDB database list}
}
\value{
Character vector of HMDB IDs for extracellular metabolites
}
\description{
Get Extracellular Metabolites
}
\keyword{internal}
