% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactate-internal.R
\name{.get_gene_contributions}
\alias{.get_gene_contributions}
\title{Get Gene Contributions for Lactate Signaling}
\usage{
.get_gene_contributions(gene_scores, gene_sets)
}
\arguments{
\item{gene_scores}{Matrix of gene scores}

\item{gene_sets}{Lactate gene sets}
}
\value{
List with production and sensing gene contributions
}
\description{
Returns the contribution of each gene to production and sensing scores.
}
\keyword{internal}
