% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.get_metabolite_pathways}
\alias{.get_metabolite_pathways}
\title{Get Pathway Information for Metabolites}
\usage{
.get_metabolite_pathways(db, hmdb_ids = NULL)
}
\arguments{
\item{db}{MetalinksDB database list}

\item{hmdb_ids}{Character vector of HMDB IDs}
}
\value{
data.frame of pathway associations
}
\description{
Get Pathway Information for Metabolites
}
\keyword{internal}
