% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.get_sensing_proteins}
\alias{.get_sensing_proteins}
\title{Get Receptor/Sensing Proteins for Metabolites}
\usage{
.get_sensing_proteins(db)
}
\arguments{
\item{db}{MetalinksDB database list}
}
\value{
data.frame of sensing relationships
}
\description{
Get Receptor/Sensing Proteins for Metabolites
}
\keyword{internal}
