% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.get_transporters}
\alias{.get_transporters}
\title{Get Transporter Proteins}
\usage{
.get_transporters(db, direction = "out")
}
\arguments{
\item{db}{MetalinksDB database list}

\item{direction}{Character. "out" for secretion, "in" for uptake}
}
\value{
data.frame of transporter relationships
}
\description{
Get Transporter Proteins
}
\keyword{internal}
