% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.load_metalinksdb}
\alias{.load_metalinksdb}
\title{Load MetalinksDB Database}
\usage{
.load_metalinksdb()
}
\value{
A list containing the MetalinksDB data
}
\description{
Load MetalinksDB Database
}
\keyword{internal}
