% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.print_progress}
\alias{.print_progress}
\title{Print Progress}
\usage{
.print_progress(current, total, prefix = "Progress")
}
\arguments{
\item{current}{Current iteration}

\item{total}{Total iterations}

\item{prefix}{Prefix message}
}
\description{
Print Progress
}
\keyword{internal}
