% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactate-internal.R
\name{.run_lactate_permutation}
\alias{.run_lactate_permutation}
\title{Run Permutation Test for Lactate Communication}
\usage{
.run_lactate_permutation(
  object,
  comm_scores,
  pathway = "direct",
  n_permutations = 100,
  method = "combined",
  comm_method = "geometric",
  min_production = 0,
  min_sensing = 0,
  include_uptake = TRUE,
  use_weights = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{scMetaLink object}

\item{comm_scores}{Communication score matrix}

\item{pathway}{Character. "direct" or "indirect"}

\item{n_permutations}{Number of permutations}

\item{method}{Scoring method}

\item{min_production}{Minimum production threshold}

\item{min_sensing}{Minimum sensing threshold}

\item{include_uptake}{Include uptake in direct sensing}

\item{use_weights}{Use weights in indirect sensing}

\item{verbose}{Print progress}
}
\value{
Matrix of p-values
}
\description{
Performs permutation testing by shuffling cell type labels.
}
\keyword{internal}
