% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safe_log}
\alias{.safe_log}
\title{Safe Log Transform}
\usage{
.safe_log(x, base = 2, offset = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{base}{Log base}

\item{offset}{Offset to avoid log(0)}
}
\description{
Safe Log Transform
}
\keyword{internal}
