% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.scale_01}
\alias{.scale_01}
\title{Scale Matrix to 0-1}
\usage{
.scale_01(mat, by_row = TRUE)
}
\arguments{
\item{mat}{Matrix}

\item{by_row}{Logical. Scale by row}
}
\description{
Scale Matrix to 0-1
}
\keyword{internal}
