% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_input}
\alias{.validate_input}
\title{Validate Input Data}
\usage{
.validate_input(expr_data, cell_meta)
}
\arguments{
\item{expr_data}{Expression matrix}

\item{cell_meta}{Cell metadata}
}
\description{
Validate Input Data
}
\keyword{internal}
