% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.zscore}
\alias{.zscore}
\title{Z-score Normalization}
\usage{
.zscore(mat, by_row = TRUE)
}
\arguments{
\item{mat}{Matrix}

\item{by_row}{Logical. Normalize by row}
}
\description{
Z-score Normalization
}
\keyword{internal}
