% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportResults}
\alias{exportResults}
\title{Export Results to CSV}
\usage{
exportResults(object, output_dir = ".", prefix = "scMetaLink")
}
\arguments{
\item{object}{scMetaLink object}

\item{output_dir}{Character. Output directory}

\item{prefix}{Character. File prefix}
}
\value{
Invisibly returns file paths
}
\description{
Export Results to CSV
}
