% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{getCommunicationMatrix}
\alias{getCommunicationMatrix}
\title{Get Communication Summary Matrix}
\usage{
getCommunicationMatrix(object, aggregate_method = "sum")
}
\arguments{
\item{object}{scMetaLink object}

\item{aggregate_method}{Character. How to aggregate metabolites}
}
\value{
Matrix of aggregated communication scores
}
\description{
Get Communication Summary Matrix
}
