% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactate.R
\name{getLactateSignalingSummary}
\alias{getLactateSignalingSummary}
\title{Get Lactate Signaling Summary}
\usage{
getLactateSignalingSummary(object, pathway = "combined", top_n = 10)
}
\arguments{
\item{object}{scMetaLink object with lactate signaling results}

\item{pathway}{Character. "direct", "indirect", or "combined". Default "combined".}

\item{top_n}{Integer. Number of top interactions to return. Default 10.}
}
\value{
data.frame summarizing top cell-cell communications
}
\description{
Returns a summary of lactate-mediated communication between cell types.
}
\examples{
\donttest{
data(crc_example)
obj <- createScMetaLink(crc_expr, crc_meta, "cell_type")
obj <- inferLactateSignaling(obj)
getLactateSignalingSummary(obj)
}
}
