% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensing.R
\name{getMetaboliteReceptors}
\alias{getMetaboliteReceptors}
\title{Get Receptors for a Metabolite}
\usage{
getMetaboliteReceptors(metabolite, include_transporters = TRUE)
}
\arguments{
\item{metabolite}{Character. Metabolite HMDB ID or name}

\item{include_transporters}{Logical. Include transporters}
}
\value{
data.frame with receptor information
}
\description{
Get Receptors for a Metabolite
}
