% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{getPathwayCommunicationMatrix}
\alias{getPathwayCommunicationMatrix}
\title{Get Pathway Communication Matrix}
\usage{
getPathwayCommunicationMatrix(object, pathway)
}
\arguments{
\item{object}{scMetaLink object}

\item{pathway}{Character. Pathway name}
}
\value{
Matrix of pathway-specific communication
}
\description{
Get Pathway Communication Matrix
}
