% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{getPathwayMetabolites}
\alias{getPathwayMetabolites}
\title{Get Metabolites in Pathway}
\usage{
getPathwayMetabolites(pathway, only_signaling = FALSE)
}
\arguments{
\item{pathway}{Character. Pathway name (can be partial match)}

\item{only_signaling}{Logical. Only return metabolites with receptors}
}
\value{
data.frame with metabolites in the pathway
}
\description{
Get Metabolites in Pathway
}
