% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getSpatialDistanceStats}
\alias{getSpatialDistanceStats}
\title{Get Spatial Distance Statistics}
\usage{
getSpatialDistanceStats(object)
}
\arguments{
\item{object}{A scMetaLink object with spatial information}
}
\value{
A list with distance statistics
}
\description{
Calculate statistics about spatial distances in the dataset
}
