% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{getSummaryStats}
\alias{getSummaryStats}
\title{Get Summary Statistics}
\usage{
getSummaryStats(object)
}
\arguments{
\item{object}{scMetaLink object}
}
\value{
list with summary statistics
}
\description{
Get summary statistics of scMetaLink analysis
}
