% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{getTopProducers}
\alias{getTopProducers}
\title{Get Top Producing Cell Types for a Metabolite}
\usage{
getTopProducers(object, metabolite, top_n = 5)
}
\arguments{
\item{object}{scMetaLink object}

\item{metabolite}{Character. Metabolite HMDB ID or name}

\item{top_n}{Integer. Number of top cell types to return}
}
\value{
data.frame with top producing cell types
}
\description{
Get Top Producing Cell Types for a Metabolite
}
