% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensing.R
\name{getTopSensors}
\alias{getTopSensors}
\title{Get Top Sensing Cell Types for a Metabolite}
\usage{
getTopSensors(object, metabolite, top_n = 5)
}
\arguments{
\item{object}{scMetaLink object}

\item{metabolite}{Character. Metabolite HMDB ID or name}

\item{top_n}{Integer. Number of top cell types to return}
}
\value{
data.frame with top sensing cell types
}
\description{
Get Top Sensing Cell Types for a Metabolite
}
