% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{listGenes}
\alias{listGenes}
\title{List Available Genes}
\usage{
listGenes(role = "all")
}
\arguments{
\item{role}{Character. Filter by role: "all", "enzyme", "receptor", "transporter"}
}
\value{
data.frame with gene information
}
\description{
Get a list of all genes in the database with their roles
}
\examples{
\donttest{
# Get all genes
genes <- listGenes()

# Get only receptor genes
receptors <- listGenes(role = "receptor")
}
}
