% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{listMetabolites}
\alias{listMetabolites}
\title{List Available Metabolites}
\usage{
listMetabolites(type = "all")
}
\arguments{
\item{type}{Character. Filter by interaction type: "all", "signaling" (lr only),
or "metabolic" (pd only)}
}
\value{
data.frame with metabolite information
}
\description{
Get a list of all metabolites in the database with their properties
}
\examples{
\donttest{
# Get all metabolites
mets <- listMetabolites()
head(mets)

# Get only signaling metabolites (those with receptors)
signaling_mets <- listMetabolites(type = "signaling")
}
}
