% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{listTopPathways}
\alias{listTopPathways}
\title{List Top Pathways}
\usage{
listTopPathways(object, n = 20)
}
\arguments{
\item{object}{scMetaLink object}

\item{n}{Integer. Number of top pathways}
}
\value{
data.frame with top pathways
}
\description{
List Top Pathways
}
