% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadScMetaLink}
\alias{loadScMetaLink}
\title{Load scMetaLink Object}
\usage{
loadScMetaLink(file)
}
\arguments{
\item{file}{Character. File path}
}
\value{
scMetaLink object
}
\description{
Load scMetaLink Object
}
