% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMetaboliteProfile}
\alias{plotMetaboliteProfile}
\title{Plot Metabolite Profile}
\usage{
plotMetaboliteProfile(object, metabolite, show_genes = FALSE)
}
\arguments{
\item{object}{scMetaLink object}

\item{metabolite}{Character. Metabolite HMDB ID or name}

\item{show_genes}{Logical. Show contributing genes}
}
\value{
A ggplot object
}
\description{
Visualize production and sensing profiles for a metabolite
}
