% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPathwayCommunication}
\alias{plotPathwayCommunication}
\title{Plot Pathway Communication}
\usage{
plotPathwayCommunication(object, top_pathways = 20, type = "bar")
}
\arguments{
\item{object}{scMetaLink object}

\item{top_pathways}{Integer. Number of top pathways to show}

\item{type}{Character. "heatmap" or "bar"}
}
\value{
A ggplot or ComplexHeatmap object
}
\description{
Visualize pathway-level communication
}
