% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTopInteractions}
\alias{plotTopInteractions}
\title{Plot Top Interactions}
\usage{
plotTopInteractions(object, top_n = 30, group_by = "metabolite")
}
\arguments{
\item{object}{scMetaLink object}

\item{top_n}{Integer. Number of top interactions}

\item{group_by}{Character. Group by "sender", "receiver", or "metabolite"}
}
\value{
A ggplot object
}
\description{
Dot plot of top metabolite-mediated interactions
}
