% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{runScMetaLinkSeurat}
\alias{runScMetaLinkSeurat}
\title{Run scMetaLink from Seurat Object}
\usage{
runScMetaLinkSeurat(
  seurat_obj,
  cell_type_column = "cell_type",
  assay = "RNA",
  slot = "data",
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{cell_type_column}{Character. Column in meta.data for cell type}

\item{assay}{Character. Assay to use}

\item{slot}{Character. Slot to use ("data" or "counts")}

\item{...}{Additional arguments passed to runScMetaLink}
}
\value{
A scMetaLink object
}
\description{
Convenience function to run analysis from Seurat object
}
