% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveScMetaLink}
\alias{saveScMetaLink}
\title{Save scMetaLink Object}
\usage{
saveScMetaLink(object, file)
}
\arguments{
\item{object}{scMetaLink object}

\item{file}{Character. File path (RDS format)}
}
\value{
Invisibly returns file path
}
\description{
Save scMetaLink Object
}
