% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{scMetaLink-class}
\alias{scMetaLink-class}
\title{scMetaLink Class Definition}
\description{
S4 class for storing metabolite-mediated cell communication analysis results
}
\section{Slots}{

\describe{
\item{\code{expression_data}}{Matrix. Normalized gene expression matrix (genes x cells)}

\item{\code{cell_meta}}{data.frame. Cell metadata with cell type annotations}

\item{\code{cell_type_column}}{Character. Column name for cell type in cell_meta}

\item{\code{production_scores}}{Matrix. Metabolite production potential scores (metabolites x cell_types)}

\item{\code{sensing_scores}}{Matrix. Metabolite sensing capability scores (metabolites x cell_types)}

\item{\code{communication_scores}}{Array. Communication scores (sender x receiver x metabolite)}

\item{\code{communication_pvalues}}{Array. P-values from permutation test}

\item{\code{significant_interactions}}{data.frame. Filtered significant interactions}

\item{\code{pathway_aggregated}}{data.frame. Pathway-level aggregated results}

\item{\code{parameters}}{list. Analysis parameters}

\item{\code{database}}{list. MetalinksDB reference data}
}}

