% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{searchGene}
\alias{searchGene}
\title{Search Gene in Database}
\usage{
searchGene(query)
}
\arguments{
\item{query}{Character. Gene symbol}
}
\value{
data.frame with gene information and associated metabolites
}
\description{
Search Gene in Database
}
