% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{searchMetabolite}
\alias{searchMetabolite}
\title{Search Metabolite in Database}
\usage{
searchMetabolite(query, exact = FALSE)
}
\arguments{
\item{query}{Character. Search query (name or HMDB ID)}

\item{exact}{Logical. Exact match only}
}
\value{
data.frame with matching metabolites
}
\description{
Search Metabolite in Database
}
