% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,scMetaLink-method}
\alias{show,scMetaLink-method}
\title{Show Method for scMetaLink}
\usage{
\S4method{show}{scMetaLink}(object)
}
\arguments{
\item{object}{scMetaLink object}
}
\description{
Show Method for scMetaLink
}
