% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial}
\alias{spatial}
\title{scMetaLink Spatial Transcriptomics Module}
\description{
Functions for analyzing metabolite-mediated cell communication
  in spatial transcriptomics data.
}
\keyword{internal}
