% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{summarizeCommunicationPairs}
\alias{summarizeCommunicationPairs}
\title{Summarize Communication by Cell Type Pairs}
\usage{
summarizeCommunicationPairs(object, aggregate_method = "sum")
}
\arguments{
\item{object}{scMetaLink object}

\item{aggregate_method}{Character. "sum", "mean", or "count"}
}
\value{
data.frame with summarized communication
}
\description{
Summarize Communication by Cell Type Pairs
}
