% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{summarizePathwayActivity}
\alias{summarizePathwayActivity}
\title{Summarize Pathway Activity}
\usage{
summarizePathwayActivity(object)
}
\arguments{
\item{object}{scMetaLink object}
}
\value{
data.frame with pathway activity summary
}
\description{
Summarize Pathway Activity
}
