% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{.check_dependencies}
\title{scMetaLink Utility Functions}
\usage{
.check_dependencies(packages, type = "error")
}
\arguments{
\item{packages}{Character vector of package names}

\item{type}{Character. "error" to stop, "warning" to warn}
}
\description{
Internal utility functions for scMetaLink package
}
\keyword{Check}
\keyword{Dependencies}
\keyword{Package}
\keyword{internal}
